refactor(docs): move useImageSelect to common components
