+++ apps/insight-viewer-docs/components/ImageSelect/index.ts
@@ -1,0 +1,1 @@
+export { default as useImageSelect, type OnSelect } from './useImageSelect'
