+++ apps/insight-viewer-docs/containers/Basic/index.tsx
@@ -1,33 +1,11 @@
-import { Box, Heading } from '@chakra-ui/react'
-import InsightViewer, { useImage } from '@lunit/insight-viewer'
-import CodeBlock from '../../components/CodeBlock'
-import { ViewerWrapper } from '../../components/Wrapper'
-import ImageSelectableViewer from './ImageSelectableViewer'
-import { DICOM_CODE, WEB_CODE } from './Code'
-import { CODE_SANDBOX } from '../../const'
-
-export const WEB_URL =
-  'https://rawgit.com/cornerstonejs/cornerstoneWebImageLoader/master/examples/Renal_Cell_Carcinoma.jpg'
+import DicomImageViewer from './DicomImageViewer'
+import WebImageViewer from './WebImageViewer'
 
 function Basic(): JSX.Element {
-  const { image } = useImage({
-    web: WEB_URL,
-  })
-
   return (
     <>
-      <ImageSelectableViewer />
-      <Box>
-        <CodeBlock code={DICOM_CODE} codeSandbox={CODE_SANDBOX.basic} />
-      </Box>
-
-      <Box mt={10} mb={6}>
-        <Heading as="h3">WebImageViewer</Heading>
-      </Box>
-      <ViewerWrapper>
-        <InsightViewer image={image} />
-      </ViewerWrapper>
-      <CodeBlock code={WEB_CODE} />
+      <DicomImageViewer />
+      <WebImageViewer />
     </>
   )
 }
